+++ src/canvas/widgets/process_table.rs
@@ -5,7 +5,6 @@ use tui::{
     text::{Line, Span},
     widgets::Paragraph,
 };
-use unicode_segmentation::UnicodeSegmentation;
 
 use crate::{
     app::{App, AppSearchState},
@@ -110,16 +109,16 @@ impl Painter {
             search_state: &AppSearchState, available_width: usize, is_on_widget: bool,
             currently_selected_text_style: Style, text_style: Style,
         ) -> Vec<Span<'_>> {
-            let start_index = search_state.display_start_char_index;
-            let cursor_index = search_state.grapheme_cursor.cur_cursor();
+            let start_index = search_state.input_field_state.display_start_index();
+            let cursor_index = search_state.input_field_state.cursor_index();
             let mut current_width = 0;
-            let query = search_state.current_search_query.as_str();
+            let query = search_state.input_field_state.current_query();
 
             if is_on_widget {
                 let mut res = Vec::with_capacity(available_width);
-                for ((index, grapheme), lengths) in
-                    UnicodeSegmentation::grapheme_indices(query, true)
-                        .zip(search_state.size_mappings.values())
+
+                for (index, grapheme, lengths) in
+                    search_state.input_field_state.graphemes_with_ranges()
                 {
                     if index < start_index {
                         continue;
@@ -145,7 +144,6 @@ impl Painter {
             } else {
                 // This is easier - we just need to get a range of graphemes, rather than
                 // dealing with possibly inserting a cursor (as none is shown!)
-
                 vec![Span::styled(query.to_string(), text_style)]
             }
         }
@@ -171,6 +169,7 @@ impl Painter {
             proc_widget_state
                 .proc_search
                 .search_state
+                .input_field_state
                 .get_start_position(available_width, app_state.is_force_redraw);
 
             // TODO: [CURSOR] blinking cursor?